# Hospital_backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hospital API Documentation",
        "description": "Documentation for Hospital Management System API",
        "contact": {
            "email": "admin@hospital.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login user",
                "description": "Authenticate user and return a token",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@hospital.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user",
                "description": "Revoke the authentication token",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get authenticated user",
                "description": "Return user profile",
                "operationId": "961ad57987905b44ae2e2f40b82e76de",
                "responses": {
                    "200": {
                        "description": "User profile retrieved"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/medecins": {
            "get": {
                "tags": [
                    "Doctors"
                ],
                "summary": "List all doctors",
                "description": "Retrieve a list of all doctors",
                "operationId": "1dadbc370c1b7f3c22f82fa290a947df",
                "responses": {
                    "200": {
                        "description": "List of doctors"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Create a new doctor",
                "description": "Store a newly created resource in storage.",
                "operationId": "d42e4bcd30b102942e11a00fa23194ee",
                "responses": {
                    "201": {
                        "description": "Doctor created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/medecins/{id}": {
            "get": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Get doctor details",
                "description": "Display the specified resource.",
                "operationId": "9a98828454615a23dccbe4286fe4e884",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Doctor details"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Update doctor details",
                "description": "Update the specified resource in storage.",
                "operationId": "ac60f416fcf2d9440132086386f44ec9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Doctor updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Delete a doctor",
                "description": "Remove the specified resource from storage.",
                "operationId": "f615485cbbdcf0ce1a2f26e3a40b603f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Doctor deleted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/rendez-vous": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "List all appointments",
                "description": "Retrieve a list of appointments with optional filtering",
                "operationId": "1280f5565a41e2068f127a52c040212d",
                "parameters": [
                    {
                        "name": "statut",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "prevu",
                                "confirme",
                                "en_cours",
                                "effectue",
                                "annule",
                                "reporte"
                            ]
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Filter by date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "medecin_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "patient_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of appointments"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Book a new appointment",
                "description": "Create a new appointment. **Note:** A confirmation email is automatically sent to the patient upon success.",
                "operationId": "10203a1b29f6a060e7b740c1eecd32c5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "patient_id",
                                    "medecin_id",
                                    "date_heure",
                                    "motif",
                                    "statut"
                                ],
                                "properties": {
                                    "patient_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "medecin_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "medecin_remplacant_id": {
                                        "description": "Optional substitute doctor ID",
                                        "type": "integer",
                                        "example": 3,
                                        "nullable": true
                                    },
                                    "date_heure": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-12-25 10:00:00"
                                    },
                                    "motif": {
                                        "type": "string",
                                        "example": "Routine Checkup"
                                    },
                                    "statut": {
                                        "type": "string",
                                        "enum": [
                                            "prevu",
                                            "confirme"
                                        ],
                                        "example": "prevu"
                                    },
                                    "observation": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Appointment created and email sent"
                    },
                    "422": {
                        "description": "Validation Error (e.g. duplicate slot)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/rendez-vous/{id}": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Get appointment details",
                "description": "Display the specified resource.",
                "operationId": "0515ec0a2650ec74fe62a14e3d80d590",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Appointment details"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Update appointment details",
                "description": "Update the specified resource in storage.",
                "operationId": "382c01a7396ddbf08dd77969e7b3ff51",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Appointment updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Cancel/Delete an appointment",
                "description": "Remove the specified resource from storage.",
                "operationId": "e33c033771e2729d45b4ec4db188a9e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Appointment cancelled successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "Doctors",
            "description": "Doctors"
        },
        {
            "name": "Appointments",
            "description": "Appointments"
        }
    ]
}